# apps/api/swag/openapi3.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Vyzorix Update Server API",
    "version": "0.0.01",
    "description": "Go-generated OpenAPI contract from handler annotations"
  },
  "paths": {
...
    "/organizations": {
      "get": {
        "summary": "",
        "description": "",
        "tags": [
          "organizations"
        ],
        "parameters": [
          {
            "name": "X-Organization-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Organization ID"
          }
        ],
        "responses": {}
      },
      "post": {
        "summary": "",
        "description": "",
        "tags": [
          "organizations"
        ],
        "parameters": [
          {
            "name": "X-Organization-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Organization ID"
          }
        ],
        "responses": {}
      }
    },
    "/organizations/{id}": {
      "get": {
        "summary": "",
        "description": "",
        "tags": [
          "organizations"
        ],
        "parameters": [
          {
            "name": "X-Organization-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Organization ID"
          }
        ],
        "responses": {}
      },
      "patch": {
        "summary": "",
        "description": "",
        "tags": [
          "organizations"
        ],
        "parameters": [
          {
            "name": "X-Organization-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Organization ID"
          }
        ],
        "responses": {}
      }
    },
...
  },
...
}